// src/types/models/IStudent.ts
export interface IStudent {
  id: number
  username?: string
  parent_name: string
  parent_phone_number: string
  description: string
  first_name: string
  last_name: string
  email: string
  birth_date: string
  gender: string
  start_date: string
  attendances_count?: number
  end_date?: string
  avatar?: string
}